Extract ListCacheKey and ListStorageData to dedicated file
- Move ListCacheKey and ListStorageData interfaces from ListQuery.ts to new ListCacheKey.ts
- Update imports across the codebase to reference new file location
- Improve code organization and reduce coupling between modules
- Maintain consistent inheritance from CollectionStorageData
